+++ packages/URLFinder/src/common/constants/patterns.ts
@@ -1,0 +1,2 @@
+// ホームページ判定用のファイル拡張子パターン
+export const HOMEPAGE_FILE_PATTERNS = ['index.html', 'index.htm', 'index.php', 'home/', 'home'];
