+++ packages/URLFinder/src/common/network/validation.ts
