--- packages/URLFinder/src/data/constants/index.ts
@@ -1,66 +1,0 @@
-/**
- * Constants統合エクスポート
- * 全ての定数への統一アクセスポイント
- */
-
-// Keywords
-export {
-  HIGH_PRIORITY_CONTACT_KEYWORDS,
-  CONTACT_KEYWORDS
-} from './keywords';
-
-// Contact Patterns
-export {
-  PHONE_PATTERNS,
-  EMAIL_PATTERN,
-  NEGATIVE_PATTERNS,
-  FORM_MENTION_PATTERNS,
-  DYNAMIC_CONTACT_KEYWORDS,
-  DYNAMIC_INDUCTION_PHRASES,
-  DYNAMIC_FORM_HINTS
-} from './contactPatterns';
-
-// HTML Analysis Keywords
-export {
-  HTML_HIGH_PRIORITY_CONTACT_KEYWORDS,
-  HTML_MEDIUM_PRIORITY_CONTACT_KEYWORDS,
-  HTML_EXCLUDED_KEYWORDS,
-  ANCHOR_SECTION_CONTACT_KEYWORDS,
-  FORM_LINK_NEGATIVE_KEYWORDS
-} from './htmlKeywords';
-
-// HTML Patterns
-export {
-  CONTACT_URL_PATTERNS,
-  NAVIGATION_SELECTORS,
-  INVALID_PAGE_PATTERNS,
-  ANCHOR_SECTION_PATTERNS,
-  FORM_LINK_PATTERNS,
-  FORM_TEXT_PATTERNS
-} from './htmlPatterns';
-
-// Form Constants
-export {
-  SUBMIT_BUTTON_KEYWORDS,
-  RECAPTCHA_PATTERNS,
-  FORM_ELEMENT_KEYWORDS,
-  SUBMIT_BUTTON_PATTERNS,
-  GOOGLE_FORMS_PATTERNS,
-  GOOGLE_FORM_URL_PATTERNS,
-  CONTACT_KEYWORDS as FORM_CONTACT_KEYWORDS,
-  CONTACT_FIELD_PATTERNS
-} from './formConstants';
-
-// URL Patterns
-export {
-  HIGH_PRIORITY_PATTERNS,
-  HIGH_CONFIDENCE_PATTERNS,
-  MEDIUM_CONFIDENCE_PATTERNS,
-  HOMEPAGE_FILE_PATTERNS
-} from './patterns';
-
-// Network
-export {
-  SNS_PATTERNS,
-  HTTP_ERROR_MESSAGES
-} from './network';
