+++ packages/URLFinder/src/flows/03_fallback/index.ts
@@ -1,0 +1,26 @@
+import type {  StrategyResult } from '../../common/types';
+import type { SearchStateData } from '../../common/types';
+import { getFinalResult } from '../../common/state';
+import { evaluateFallbackUrlQuality } from './utils';
+
+/**
+ * Fallback検索戦略
+ * SearchStateに蓄積された情報を利用してフォールバック結果を提供
+ */
+export const fallbackSearch = (baseUrl: string, searchState: SearchStateData): StrategyResult => {
+  const fallbackResult = getFinalResult(searchState);
+
+  if (fallbackResult.contactUrl && fallbackResult.foundKeywords.length > 0) {
+    const pattern = fallbackResult.foundKeywords.find(k => k.startsWith('/')) || '';
+    const qualityScore = evaluateFallbackUrlQuality(fallbackResult.contactUrl, pattern);
+
+    fallbackResult.foundKeywords.push(...qualityScore.keywords);
+    if (fallbackResult.searchMethod === 'final_fallback') {
+        fallbackResult.searchMethod = qualityScore.confidence >= 0.7 ? 'final_fallback_high_confidence' : 'final_fallback_low_confidence';
+    }
+
+    return { result: fallbackResult, newState: searchState };
+  }
+
+  return { result: null, newState: searchState };
+};
