+++ packages/URLFinder/src/flows/index.ts
@@ -1,14 +1,14 @@
 /**
  * 📋 URLFinder フロー統合エクスポート
- * 
+ *
  * 全検索フローを一括インポート可能にするバレル輸出
  * フォルダベース構造で可読性向上、import の複雑性解消
  */
 
 // 前処理フロー（preprocessing/index.ts）
-export { snsCheck, domainCheck } from './preprocessing';
+export { snsCheck, domainCheck } from './00_preprocessing';
 
 // 検索戦略フロー（各フォルダ/index.ts）
-export { urlPatternSearch } from './urlPattern';
-export { htmlAnalysis } from './htmlAnalysis';
-export { fallbackSearch } from './fallback';
+export { urlPatternSearch } from './01_urlPattern';
+export { htmlAnalysisSearch as htmlAnalysis } from './02_htmlAnalysis';
+export { fallbackSearch } from './03_fallback';
